# docs/swagger.yaml
definitions:
...
  constants.Status:
    enum:
    - 1
    - 2
    - 3
    - 9
    type: integer
    x-enum-varnames:
    - StatusAtivo
    - StatusInativo
    - StatusBloqueado
    - StatusCancelado
...
  dto.GrupoUsuarioListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.GrupoUsuarioResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.GrupoUsuarioRequest:
    properties:
      created_by:
        type: integer
      descricao:
        maxLength: 100
        type: string
      empresa_filial_id:
        type: integer
      situacao:
        type: integer
      updated_by:
        type: integer
    required:
    - descricao
    - empresa_filial_id
    type: object
  dto.GrupoUsuarioResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      descricao:
        type: string
      empresa_filial_id:
        type: integer
      empresa_filial_nome:
        type: string
      id:
        type: integer
      situacao:
        $ref: '#/definitions/constants.Status'
      situacao_label:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /grupos-usuario:
    get:
      description: Retorna uma lista paginada de grupos de usuário
      parameters:
      - default: 10
        description: Limite de itens por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Número de itens a pular
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GrupoUsuarioListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Lista todos os grupos de usuário
      tags:
      - Grupo de Usuário
    post:
      consumes:
      - application/json
      description: Cria um novo grupo de usuário com os dados fornecidos
      parameters:
      - description: Dados do Grupo de Usuário
        in: body
        name: grupo
        required: true
        schema:
          $ref: '#/definitions/dto.GrupoUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GrupoUsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cria um novo grupo de usuário
      tags:
      - Grupo de Usuário
  /grupos-usuario/{id}:
    delete:
      description: Exclui logicamente um grupo de usuário pelo ID
      parameters:
      - description: ID do Grupo de Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Exclui um grupo de usuário
      tags:
      - Grupo de Usuário
    get:
      description: Retorna um grupo de usuário específico pelo seu ID
      parameters:
      - description: ID do Grupo de Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GrupoUsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Busca um grupo de usuário pelo ID
      tags:
      - Grupo de Usuário
    put:
      consumes:
      - application/json
      description: Atualiza um grupo de usuário com os dados fornecidos pelo ID
      parameters:
      - description: ID do Grupo de Usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do Grupo de Usuário para atualização
        in: body
        name: grupo
        required: true
        schema:
          $ref: '#/definitions/dto.GrupoUsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GrupoUsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Atualiza um grupo de usuário existente
      tags:
      - Grupo de Usuário
...
swagger: "2.0"